feat: Move model code to modelstore mod
