+++ src/app.rs
@@ -1,3 +1,4 @@
+
 #[derive(Clone)]
 pub enum Format {
     JSON,
